+++ lib/mixins/pointer_detector.dart
@@ -72,7 +72,7 @@ mixin PointerDetectorHandler on Component {
   bool hasGesture() {
     if (this is DragGesture && (this as DragGesture).enableDrag) return true;
     if (this is TapGesture && (this as TapGesture).enableTab) return true;
-    if (this is MouseGesture && (this as MouseGesture).enableMouseGesture) {
+    if (this is MouseListener && (this as MouseListener).enableMouseGesture) {
       return true;
     }
     return false;