+++ example/lib/pages/mini_games/platform/fox_player.dart
@@ -40,7 +40,7 @@ class FoxPlayer extends PlatformPlayer with HandleForces {
 
   @override
   bool onBlockMovement(Set<Vector2> intersectionPoints, GameComponent other) {
-    if (other is TileWithCollision && other.type == 'tree_trunk') {
+    if (other is TileWithCollision && other.tileClass == 'tree_trunk') {
       if (jumpingState == JumpingStateEnum.up) {
         inTrunk = true;
       } else if (other.top > center.y) {