+++ example/lib/pages/player/simple/human.dart
@@ -1,7 +1,7 @@
 import 'package:bonfire/bonfire.dart';
 import 'package:example/pages/player/simple/human_sprite_sheet.dart';
 
-class HumanPlayer extends SimplePlayer {
+class HumanPlayer extends SimplePlayer with TileRecognizer {
   HumanPlayer({
     required Vector2 position,
   }) : super(