+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Next
 - Fix `RotationPlayer` animation bug.
 - Add an option for using to enable diagonal input events on `KeyboardConfig` and `JoystickDirectional`.
+- `BonfireCamera` Improvements
+- reorganized the input classes
 
 
 # 3.0.0-beta.15